refactor: carStatus 형태변환 로직 파일화 및 불필요 주석제거
