+++ src/routers/carRouter.ts
@@ -11,7 +11,6 @@ import {
 import { asyncHandler } from '../lib/async-handler';
 import upload from '../middlewares/uploadMiddleware';
 import { verifyAccessToken } from '../middlewares/verifyAccessToken';
-import { AuthenticatedRequest } from '../types/express';
 
 const carRouter = express.Router();
